review: Move ExtensionModule from sdk to statefun-flink-extensions
