+++ statefun-flink/statefun-flink-core/src/main/java/org/apache/flink/statefun/flink/core/httpfn/TransportClientsModule.java
@@ -20,7 +20,7 @@ package org.apache.flink.statefun.flink.core.httpfn;
 
 import com.google.auto.service.AutoService;
 import java.util.Map;
-import org.apache.flink.statefun.sdk.spi.ExtensionModule;
+import org.apache.flink.statefun.extensions.ExtensionModule;
 
 @AutoService(ExtensionModule.class)
 public class TransportClientsModule implements ExtensionModule {
