+++ statefun-flink/statefun-flink-core/src/main/java/org/apache/flink/statefun/flink/core/spi/ExtensionResolver.java
@@ -18,8 +18,8 @@
 
 package org.apache.flink.statefun.flink.core.spi;
 
+import org.apache.flink.statefun.extensions.ExtensionModule;
 import org.apache.flink.statefun.sdk.TypeName;
-import org.apache.flink.statefun.sdk.spi.ExtensionModule;
 
 /**
  * Resolves a bound extension (bound by {@link ExtensionModule}s) given specified {@link TypeName}s.
